flask_dashboard/requirements.txt
Flask
paho-mqtt
requests
joblib
pandas
numpy
scikit-learn
plotly
